Adds slash-commands components
GuildID is now defined based on the presence of an environment
variable. This has been added so that slash-commands can be tested
as a "Guild command", as opposed to a Global command. The difference
here is that the commands are exclusive to a Guild, and can be tested
quickly before publishing them properly. The Readme has been updated to
reflect this.

"guidefetch" slash command has been added in a very rough-cut manner. No
structs, or anything fancy for now - just getting something that works.
This will be part of the first feature for Ralphbot.

Dockerfile has been updated so that `go build` builds the project and
its internal modules. This has been added since "guidefetch" is part of
an internal library, the path needs to be added so the project builds.

"client" folder has been renamed to "ralphbot" to better match the
envisioned project structure.