Adds helper scripts for secrets and ssm params
These scripts provide the means to publish AWS SSM Parameters, and AWS
Secrets Manager Secrets to an AWS Account that ralphbot would be
deployed to. In doing so, these secrets are not published in source
code in plaintext, and are only retrieved internally via AWS
CloudFormation, or other appropriate AWS Services.