Adds Infrastructure Stack for ralphbot (#11)
* Adds prettier, eslint, and rimraf

- `prettier` provides styling. This can be leveraged in CI to enforce
code styling for the CDK Stack and its accompanying files.
https://www.npmjs.com/package/prettier

- `eslint` provides linting, fixing up problems. This can integrate with
prettier as well.
https://www.npmjs.com/package/eslint

- `rimraf` has been added to provide a means to os-agnostically
cleanup files left behind by tsc during linting, or builds where
`--clean` wasn't used. If there is a better way of doing this when
running eslint; I don't know about it.
https://www.npmjs.com/package/rimraf

* Expands CDK Stack

The boilerplate CDK stack source code has been modified (to put it
lightly) to have the necessary constructs, and resources to deploy
ralphbot to AWS.

* Adds helper scripts for secrets and ssm params

These scripts provide the means to publish AWS SSM Parameters, and AWS
Secrets Manager Secrets to an AWS Account that ralphbot would be
deployed to. In doing so, these secrets are not published in source
code in plaintext, and are only retrieved internally via AWS
CloudFormation, or other appropriate AWS Services.