ci: Reconfigures Github Actions
- Golang tests have been removed: A proper test suite will be added in
the future. A successful build doesn't represent adequete testing.
- Infrastructure stack tests are now their own workflow: This is to
better seperate duties of the workflows, and makes it easier to
maintain.
- Workflow trigger changed to `on push`: When tests are executed, they
now should run without the need for a PR to be raised. Once a branch
exists on `origin` and a push has been completed, the configured actions
will run.