test: Adds snapshot tests to infrastructure stacks
ralphbot's infrastructure stack now includes snapshot tests using jest.

These will be critical to provide greater confidence in dependency
management to ensure that changes are detected, and approriately
investigated in the event of a dependency update.

Additionally, these snapshot tests will ensure that the stacks do not
change unintentionally as ralphbot's infrastructure stack code is
maintained.

These added into the Github actions workflows so that these tests are
conducted automatically whenever changes are pushed on a branch. This
is useful because not only will test failure be signaled in an automatic
fashion, but also it will act as a blocker for deployments, preventing
upstream issues in the release process.