chore(vscode): Adds vscode extension and settings A `.vscode` directory has been added, which features extension recommendations and settings for these extensions.
`eslint` and `prettier is important for this project's infrastructure
stack's maintainability, so having these not only recommeneded, but also
running on save is a value-add to this project, rather than having to
remember to run a command before every commit.

An experimental setting concerning `gopls` has been added to. The `go`
extension for vscode behaves oddly for users running `asdf` despite
claims that this has been fixed.