chore(eslint): Removes eslintrc.js configs In eslintrc.js, there is a configuration concerning the project paths.
Although the previous configuration is a workaround for `vscode`'s
`eslint` extension, the configuration is no longer needed since `eslint`
will be running within the same directory as the project is executed as
a command. Whereas, the extension will always run for any file format
that it is compatible with. Hence, this configuration is no longer
required for the project.