feat(docker-compose): Adds docker-compose
A `docker-compose-local` file has been added to the `golang` portion of the
`ralphbot` repo.

The purpose of the `docker-compose-local` file is to provide an
interface in running `ralphbot` locally in a declarative fashion. Though
the `docker-compose` command is intended to be used alongside an
eventual `make` target following PR#85.

PR#85: https://github.com/therealvio/ralphbot/pull/85