chore(docker): Adds build argument The build argument "MAKE_TARGET" has been added to the Dockerfile.
This change allows the ability to use the MAKE_TARGET argument to
distinguish whether the explicit `make build-fargate-binary` (ARM64)
should be used, or have GOARCH defined automatically when running the
`make build` target.

The purpose of this change is to seperate what type of build occurs.
The argument shouldn't need to be passed in a local development context
unless the intention is to test the argument. It should only be used
during deployment - which is handled via CI/CD.