chore(cicd): Adds a workaround to GHA trigger The "src/**" directory has been added to the scope of the "infra" linting GHA workflow.
Due to a lack of workflows that trigger around the `golang` "src"
irectory that have their own trigger relationship with the "deploy"
workflow, this change has been added as a means to force the "deploy"
workflow to happen as the `golang` "src" directory is worked on.

This should be removed by the time `golang` code tests have been added
as a Github workflow.