refactor(cicd): Refactors ECR Auth, and registry The steps involving ECR authentication have been changed to use the `aws-actions/amazon-ecr-login@v1` Github action.
Using this workflow, the registry URL has been changed to an environment
variable called "REGISTRY" in an effort to simplify `docker` command
maintenance.
In the original implementation had used concatinations to create the
registry URL and resulted in lengthy commands. This in turn increases
the chance of making mistakes when making modifications to the workflows
that use these.

These changes should reduce the amount of effort needed to make any
modifications to the `docker` commands pointing to these URLs.