refactor(cicd): GHA Workflows (#95)
* fefactor(cicd): Renames "ECR_NAME" to "REPOSITORY"
The workflow-scope environment variable "ECR_NAME" has been renamed to
"REPOSITORY".
The intention of this rename is to clarify intention by naming the
variable to what it refers to correctly. ECR_NAME is really just the
repository name. In a later commit, the "REGISTRY" environment variable
will be used to indicate the registry, which is the AWS account's
private registry.

All in all, these two commits should clarify what means what in the
various docker commands.

* refactor(cicd): Refactors ECR Auth, and registry
The steps involving ECR authentication have been changed to use the
`aws-actions/amazon-ecr-login@v1` Github action.

Using this workflow, the registry URL has been changed to an environment
variable called "REGISTRY" in an effort to simplify `docker` command
maintenance.
In the original implementation had used concatinations to create the
registry URL and resulted in lengthy commands. This in turn increases
the chance of making mistakes when making modifications to the workflows
that use these.

These changes should reduce the amount of effort needed to make any
modifications to the `docker` commands pointing to these URLs.

* refactor(cicd): Changes "VERSION" to "IMAGE_TAG"
The use of the "VERSION" variable has been renamed to "IMAGE_TAG" where
appropriate.
The aim of this change is to clarify intention on what the variable is
supposed to refer to, and maintain consitency throughout the workflow
code.

* chore(cicd): Sorts cdk-deploy needs

* chore(cicd): Removes unused step "ids"
Unused step "ids" have been removed throughout the Github Actions.
Step ids should be used when outputs, or the step themselves are being
referred to in subsequent steps or jobs. Hence, if they're not being
referred to, then they're not being used and are just "fluff" in the
code.