chore(gitignore): Corrects path for .js files The .gitignore rule concerning .js files in the __tests__ dir has been corrected. Now when "tsc" or "yarn build" is run, .js files created in the "__tests__" directory will not accidentally be detected by git.
When running certain node commands, the .js files are read over
Typescript files. This is a standard behaviour since normally "ts" files
need to be transpiled to "js". Though with the use of "ts-node", this
can be interpreted on the fly without the need to produce the build
artifacts.

If caught unaware of this, changes to "ts" code will be ineffectual
until the "js" files are either removed, or replaced following a build.