chore(node): Removes "rimraf" "rimraf" has been removed from the infra stack packages, and associated scripts have also been replaced with "tsc --build --clean".
At some point during `tsc`'s releases, "--clean" was added. This is
intended to clean up build artifacts for "ts" project files. Being
that this is a native feature of Typescript, and is much simplier to use
than a custom `rimraf` script, there is no need for `rimraf` anymore.

So, if `yarn build` is ever run, `yarn clean` can be run afterwards to
clean up `js` files. Although building the project isn't required thanks
to `ts-node`, this is a tool to help keep the repository tidy and to
troubleshoot a likely issue faced with Typescript projects.

Some background:
When running certain node commands, the .js files are read over
Typescript files. This is a standard behaviour since normally "ts" files
need to be transpiled to "js". Though with the use of "ts-node", this
can be interpreted on the fly without the need to produce the build
artifacts.

If caught unaware of this, changes to "ts" code will be ineffectual
until the "js" files are either removed, or replaced following a build.