refactor(config): Adds config package
A `config` package has been added to the `ralphbot` source code to
replace parameter parsing.

Prior to this commit, when `ralphbot` executes, the configurations for
it were set up as flags that are defined when the executable as both
environment variables, as flags passed before execution. For
consistency's sake, these have been set up as configurations based
on environment variable definitions. This all-or-nothing approach
should increase readability, and prompote modularity.

The environment variable parsing is accomplished with `envconfig`.