chore: Changes `ralphbot` service image to alpine The container image for `ralphbot` has been changed from the `golang`-specific image to the standard `alpine` image.
Given that `ralphbot` is a `golang` app, it compiles everything it needs
into the binary during build time. This means, it isn't necessary to
have anything in the container image besides the Linux kernal and
a very light distro.

Doing this change also reduces the final `ralphbot` image to 13MB from
330MB at the time of this commit.