refactor(internal): Changes cmd funcs to private
The functions in the `discord` package have been changed to private
functions rather than being exported as public commands.

The code in the `command` package is so far only being used in a context
where it is internal to the `discord` package and hence exporting the
function isn't required at the moment.