fix: Replaces deprecated set-output call in GHA
The `::set-output` method in the deploy Github Action workflow is
replaced with a new method involving the use of the $GITHUB_OUTPUT
redirection.

In a Github Blog post, the `::set-output` method is being deprecated in
favour of redirecting outputs to a dedicated environment variable.
By 31st March 2023, the `::set-output` method will no longer work. This
change updates the workflow to move away from the old method.

Github Blog Post:
https://github.blog/changelog/2022-10-11-github-actions-deprecating-save
-state-and-set-output-commands/