feat: Adds link-dump command
A new slash-command has been added where, if invoked, will send a direct
message to the invoker. The message contains a series of links defined
in the `links.json` file. These links are for popular Destiny 2 web
apps.

This command was suggested to so that if newcomers asked for links to
tools and services we use, we can instruct them to use the command
rather than "fish" for it.