chore: removes renovate config
This change is part of an effort to force renovate to remake this repo's
renovate configuration PR.

The current config previously had issues with AWS Public ECR gallery,
and right now `renovate` is having issues with updating `go`
dependencies.

The configure renovate PR provides a lot more useful information on how
dependencies will be managed based on the configuration provided.