Enables autormerging for controllable deps - Dockerfiles aren't incredibly controllable when it comes to major versions. `Golang` does not have this problem, as the only major version is v1. Whereas, for `Nodejs`, it has multiple major versions, and ideally we'd want to control the version to LTS. While `asdf`, and `npm` support this version pinning/range behaviour, `docker` currently doesn't due to how versioning is tied to mutable tags. Not to mention, the tags themselves are not strictly SemVer.
- For now, `Nodejs` is going to be set to manual PRs until better
support for Docker versioning is available. The current image version
could be supported better if `-alpine` wasn't included in the version,
though for the sake of keeping container images lean, `-alpine` is
necessary.

- The most that is necessary when a new `nodejs` version is out is to
update the `Dockerfile` FROM image major version to match up
with the `npm` and `asdf` version. Whether it be manually updating
the version number, or merging the `nodejs` Docker PR into the
`node.js` PR, and merging them altogether.