chore: replace os.ReadFile() with go:embed This fixes an odd bug in testing within the `discord` library where the `dadjoke` jokes.json file was attempting to be loaded during test runs.
This seems to come from a daisy chain of "required" function calls
starting from the `bot.go` portion of the package. Likely due to the
`jokes` variable *needed* to be loaded as part of registering the
`dadjoke` command. It seems that golang loads more than just the
"to-be-tested" function when the wider library is called.

This fix utilizes go:embed to embed the file in during the build, and
mitigate this problem, since the embed happens in context to the
*.go file using it. Though a proper fix would be to use a struct, rather
than reading embedding a json file, though this isn't an urgent thing.