node deps within package.json can be automatically updated
`dockerfile` dependency manager still lacks support for version control,
but this may never happen for a long given how inconsistent tagging is
for images. Though when it eventually has the ability to source version
numbers from other managers as a composite type*, then we'll be in
business.

* = A Dockerfile, with `semver` versioning schemes, since NodeJS Docker
images honour their semver scheme for NodeJS.