chore: Enable automerge of golang deps
In PR #411, a bulk of tests were added that increase confidence in code
functionality when running tests.

These tests were critical in determining that dependency upgrades would
not cause issues with `ralphbot`'s operations. At this stage,
new cases would need to be added in future scenarios - for now, the
coverage is sufficient.