fix: invocations were not randomising
The `makePhrase` call was captured as a variable when the
GetCommandHandlers was invoked. Since GetCommandHandlers is only
supposed to run during startup of the bot service, the selected coin
side and phrase saved to the variable would be used.

This change tweaks the content response so that the function is executed
each time the interaction is invoked.