doc: Provide at-a-glance status for Ralphbot
This change adds Build badges for Ralphbot. This is to provide an
at-a-glance look for Ralphbot's project health.

This change is more useful for onlookers that aren't familiar with
how to run a service, but want a way to check if Ralphbot is currently
deployed or not (from main), and if health-checks are passing generally.