chore: regexManager for pnpm upgrades
Version management for the Github action `pnpm/setup-pnpm` has been
added via a regexManager rule. This is due to a limitation of the action
not having a means to specify a directory where a `package.json` can
live. This means rather than relying on the versioning of a package
manager specified in there, a version has to be set for the action.

The regexManager rule is based on a pre-existing pattern made by
renovate.

https://docs.renovatebot.com/presets-regexManagers/