chore: platformAutomerge is annoying to work with
There's a quirk where platformAutomerge can potentially merge PRs even
if tests have failed. It is recommended by `renovate` that you leverage
your Git hosting platform's capabilities to enforce test passing before
PR merge. The problem here is that Github's options for this suck since
you can only specify actions that have run in the last week. So if you
wish to specify other status checks, they had to have run recently.

platformAutomerge is enabled by default. When this is disabled,
`renovate` is supposed to use it's own automerge capabilities.

This change is intended to get around this problematic behaviour because
we have had a serious of PRs with failing tests get automerged. It is
preferred behaviour that the PRs don't get merged so that fixes for
the dependency updates are introduced into the PR itself.

Docs on platformAutomerge:
https://docs.renovatebot.com/configuration-options/#platformautomerge