ci: ralphbot now deploys using OIDC
These changes implement IAM changes within ralphbot's AWS account where
Github is now an identity provider via OIDC. Meaning, we can leverage
short-lived credentials during deployment rather than static Access
Keys.

Ref:
https://docs.github.com/en/actions/deployment/
security-hardening-your-deployments/
configuring-openid-connect-in-amazon-web-services

Notes:
- The addition of permissions, this is under "adding permissions
settings"
- When specifying the OIDC role trust policy, if you use an environment,
branch filtering in the policy will not work if the branch uses an
environment. In our case, `main` uses `production-global`. So,
instead we need to scope the policy to use the `environment` in the
subject. For more information, refer to:
https://docs.github.com/en/actions/deployment/
security-hardening-your-deployments/
about-security-hardening-with-openid-connect#
filtering-for-a-specific-branch