ci: OIDC broke CDK deploy via container
This change removes the build and deployment steps concerning the
containerisation of ralphbot's CDK app.

The migration to the OIDC-based auth process broke how credentials are
passed into Docker when running the CDK container, where the credentials
are not present. This will need to be looked at. For now, we can run
the CDK commands from the ./ops/ directory directly.