Revert "chore(git): ignore bin and pkg dirs made by golang vscode extension"
This reverts commit f3ae9551d4d5a6b54a4e06ed4b3dba0373c792d7.

The fix to this is a little odd. The behaviour was caused by the
vscode-go extension, however it appears to have been caused by it not
being able to identify the `GOROOT` environment variable.

There's 2 problems contributing to this:

1. The `GOROOT` version isn't maintained automatically when switching
between versions via asdf. Your shell config needs to reference a
script so that each time the shell loads GOROOT reflects the installed
version of `go` setup by `asdf`.
Refer to
https://github.com/asdf-community/asdf-golang/blob/33b1f6d73a408b32dee8c93b1763c0ba102ee45d/README.md?plain=1#L28-L36

Reloading the shell shouldn't be necessary as versions of `go` are
switched once this configuration is set up.

2. I previously had `golangci-lint` installed via `homebrew`, however
the problem here is that it has a dependency on specific versions of `go`
being installed via `homebrew` as well. This interferes with `asdf` in
that the homebrew-installed version of `go` is _THE_ the version of go
that is referenced when `go` is called in your shell (e.g. `which go`).

To get around this, I had to uninstall `golangci-lnt` and `go` via
`homebrew`. If I wanted to really use that tool again, I could either
use their official docker image, or build the source for myself. The
latter of which isn't recommended by the development team since there is
a strict coupling between `go` version used to build the package.