fix: no double-dipping on deploys if one config is changed
This commit fixes an issue where the deploy job would run twice if
either the ralphbot-ops-lint.yaml or ralphbot-src-lint.yaml files were
modified. Causing one deploy to fail, while another succeeds.

This triggers because the workflow for deployment is listening in on
the completion of either of these other two workflows. Triggering
the linting workflows to only occur for their target directory changes
and themselves should fix this issue.