chore: set defaults using .envrc
This change makes it so that `.envrc` is now used as the source of truth
for environment variables, and their defaults. The
`source_env_if_exists` directive has been included to allow overrides
to be applied to these default values.

This means `.envrc.example` is no longer a required file, and simplifies
how the project works.